% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dev_history.R
\name{add_dev_history}
\alias{add_dev_history}
\title{Add dev_history.Rmd file that drives package development}
\usage{
add_dev_history(pkg = ".", overwrite = FALSE, open = TRUE, dev_dir = "dev")
}
\arguments{
\item{pkg}{Path where to save file}

\item{overwrite}{Whether to overwrite existing dev_history.Rmd file}

\item{open}{Logical. Whether to open file after creation}

\item{dev_dir}{Name of directory for development Rmd files. Default to "dev".}
}
\value{
Create a dev_history.Rmd file and return its path
}
\description{
Add dev_history.Rmd file that drives package development
}
\examples{
# Create a new project
tmpdir <- tempdir()
dummypackage <- file.path(tmpdir, "dummypackage")
dir.create(dummypackage)
browseURL(dummypackage)

# Add
add_dev_history(pkg = dummypackage, open = FALSE)

# Delete dummy package
unlink(dummypackage, recursive = TRUE)
}
