% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{add_fun_code_examples}
\alias{add_fun_code_examples}
\title{Add examples in function code}
\usage{
add_fun_code_examples(parsed_tbl, fun_code)
}
\arguments{
\item{parsed_tbl}{tibble of a parsed Rmd}

\item{fun_code}{R code of functions in Rmd as character}
}
\description{
Add examples in function code
}
