% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{add_fun_to_parsed}
\alias{add_fun_to_parsed}
\title{Add function name to parsed_tbl ----}
\usage{
add_fun_to_parsed(parsed_tbl, fun_names)
}
\arguments{
\item{parsed_tbl}{tibble of a parsed Rmd}

\item{fun_names}{Names of functions in Rmd}
}
\description{
Add function name to parsed_tbl ----
}
