% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fusen_chunks.R
\name{add_fusen_chunks}
\alias{add_fusen_chunks}
\title{Add \code{{fusen}} chunks}
\usage{
add_fusen_chunks(
  function_name = NULL,
  export = getOption("fusen.export.functions")
)
}
\arguments{
\item{function_name}{Name of the function to create.
If NULL (the  default), the user will be prompted to enter it.}

\item{export}{Should the function be exported?
Default is \code{getOption("fusen.export.functions")}. If NULL, the
user will be prompted to enter it.}
}
\value{
A list with the context and the content, invisibly.
}
\description{
Create \code{{fusen}} chunks inside your Rmd
}
\examples{
add_fusen_chunks("this")
}
