% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fusen_files.R
\name{clean_fusen_files}
\alias{clean_fusen_files}
\title{Title}
\usage{
clean_fusen_files()
}
\value{
Used for side effect. Delete unregistered files.
}
\description{
Description
}
\examples{
clean_fusen_files()
}
