% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_parts.R
\name{combine_parsed}
\alias{combine_parsed}
\title{Combine a parsed tbl Rmarkdown file into a new Rmd file}
\usage{
combine_parsed(parsed_tbl)
}
\arguments{
\item{parsed_tbl}{A tibble with 6 columns as issued by \code{get_rmd_parts()}}
}
\value{
A Rmd file
}
\description{
Combine a parsed tbl Rmarkdown file into a new Rmd file
}
