% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{create_functions_all}
\alias{create_functions_all}
\title{Create function code, doc and tests ----}
\usage{
create_functions_all(parsed_tbl, fun_code, pkg)
}
\arguments{
\item{parsed_tbl}{tibble of a parsed Rmd}

\item{fun_code}{tibble as issued from \code{get_functions}}

\item{pkg}{Path to package}
}
\description{
Create function code, doc and tests ----
}
