% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{create_r_files}
\alias{create_r_files}
\title{create R file with code content and fun name}
\usage{
create_r_files(fun_code, pkg)
}
\arguments{
\item{fun_code}{R code of functions in Rmd as character}

\item{pkg}{Path to package}
}
\description{
create R file with code content and fun name
}
