% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{create_tests_files}
\alias{create_tests_files}
\title{Check if there are unit tests ----}
\usage{
create_tests_files(parsed_tbl, pkg_path)
}
\arguments{
\item{parsed_tbl}{tibble of a parsed Rmd}

\item{pkg_path}{Path to package}
}
\description{
Check if there are unit tests ----
}
