% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fusen_files.R
\name{df_to_config}
\alias{df_to_config}
\title{Add a tibble of files and types to the 'fusen' config file}
\usage{
df_to_config(df_files, flat_file_path = "keep", force = FALSE)
}
\arguments{
\item{df_files}{A dataframe with 'type' and 'path' columns
or a csv file path as issued from \verb{[check_not_registered_files()]}
or nothing (and it will take the csv file in "dev/")}

\item{flat_file_path}{Character. Usually \code{"keep"} or the name of the origin flat file, although inflating the flat file should have the same result.}

\item{force}{Logical. Whether to force writing the configuration file even with potential errors.}
}
\value{
Config file path.
Side effect: create a yaml config file.
}
\description{
Add a tibble of files and types to the 'fusen' config file
}
\details{
\itemize{
\item Only types: "R", "test", "vignette" are accepted.
\item All \code{path} should exist.
\item Change path to config file with \code{options(fusen.config_file = "dev/config_fusen.yaml")}
}
}
\examples{
# Add your own list of files to "keep",
# if they are not in a flat file.
# Otherwise, they may be deleted with your next `inflate()`
my_files_to_protect <- tibble::tribble(
  ~type, ~path,
  "R", "R/zaza.R",
  "R", "R/zozo.R",
  "test", "tests/testthat/test-zaza.R",
  "vignette", "vignettes/my-zaza-vignette.Rmd"
)

# \dontrun{
df_to_config(my_files_to_protect)
# }
}
