% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusen-package.R
\docType{package}
\name{fusen-package}
\alias{fusen}
\alias{fusen-package}
\title{fusen: Build a Package from Rmarkdown Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use Rmarkdown First method to build your package. Start your package with documentation, functions, examples and tests in the same unique file. Everything can be set from the Rmarkdown template file provided in your project, then inflated as a package. Inflating the template copies the relevant chunks and sections in the appropriate files required for package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thinkr-open.github.io/fusen/}
  \item \url{https://github.com/Thinkr-open/fusen}
}

}
\author{
\strong{Maintainer}: Sebastien Rochette \email{yohann@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
