% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflate_parameters.R
\name{get_inflate_parameters}
\alias{get_inflate_parameters}
\title{Returns the parameters used in an \code{inflate()} call in some flat Rmd file}
\usage{
get_inflate_parameters(rmd_flat_file)
}
\arguments{
\item{rmd_flat_file}{flat file to parse}
}
\value{
a tibble with 2 columns \code{parameter} and \code{value}
returns NULL if no parameter are provided in the \code{inflate()} call
}
\description{
Returns the parameters used in an \code{inflate()} call in some flat Rmd file
}
\examples{
get_inflate_parameters(rmd_flat_file = system.file("files_to_parse", "flat_file_to_parse_with_extra_param.Rmd", package = "fusen"))

get_inflate_parameters(rmd_flat_file = system.file("files_to_parse", "flat_file_to_parse_with_no_param.Rmd", package = "fusen"))

}
