% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_parts.R
\name{get_rmd_parts}
\alias{get_rmd_parts}
\title{Group lines of a Rmd or Qmd file by types in a tibble}
\usage{
get_rmd_parts(file)
}
\arguments{
\item{file}{A Rmd or Qmd file path}
}
\value{
A tibble with 6 columns:
\itemize{
\item \code{type}: type of the part (yaml, heading, inline, block)
\item \code{label}: label of the part (if any)
\item \code{params}: parameters of the part (if any)
\item \code{text}: text of the part (if any)
\item \code{code}: code of the part (if any)
\item \code{heading}: heading of the part (if any)
}
}
\description{
Group lines of a Rmd or Qmd file by types in a tibble
}
\examples{
file <- system.file("tests-templates/dev-template-parsing.Rmd",
  package = "fusen"
)
get_rmd_parts(file)
}
