% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fusen_files.R
\name{guess_flat_origin}
\alias{guess_flat_origin}
\title{Guess flat file of origin of a script}
\usage{
guess_flat_origin(path)
}
\arguments{
\item{path}{Path of the script to test}
}
\description{
Guess flat file of origin of a script
}
