% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate_all.R
\name{inflate_all}
\alias{inflate_all}
\title{Title}
\usage{
inflate_all(pkg = ".")
}
\value{
side effect. Inflates all your flat files that can be inflated.
}
\description{
Description
}
\examples{
\dontrun{
inflate_all()
}
}
