% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_share_on_github.R
\name{init_share_on_github}
\alias{init_share_on_github}
\title{Initiate GitHub to share your work on a website}
\usage{
init_share_on_github(ask = TRUE)
}
\arguments{
\item{ask}{Logical. \code{TRUE} (default) to ask the user to apply the instructions each time needed,
or \code{FALSE} if the user already know what to do.}
}
\value{
The URL of the website created
}
\description{
This uses {pkgdown} to share the documentation of the package through GitHub Actions.
You may need to run \code{usethis::create_github_token()}, then \code{gitcreds::gitcreds_set()} before.
}
\examples{
\dontrun{
init_share_on_github()
}
}
