% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_the_package.R
\name{map_the_package}
\alias{map_the_package}
\title{Map information of a 'fusen' built package for developers}
\usage{
map_the_package(yaml_fusen_file, emoji = TRUE)
}
\arguments{
\item{yaml_fusen_file}{The path to the yaml file
to get the map from}

\item{emoji}{Add emojis to the output}
}
\value{
A list of information about the package
}
\description{
Map information of a 'fusen' built package for developers
}
\examples{
\dontrun{
# This only works inside a 'fusen' built package
map <- map_the_package()
draw_the_map(map)
}

# Example with a dummy package
dummypackage <- tempfile("drawmap")
dir.create(dummypackage)

# {fusen} steps
fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
dev_file <- suppressMessages(
  add_flat_template(pkg = dummypackage, overwrite = TRUE, open = FALSE)
)
flat_file <- dev_file[grepl("flat_", dev_file)]

usethis::with_project(dummypackage, {
  suppressMessages(
    inflate(
      pkg = dummypackage, flat_file = flat_file,
      vignette_name = "Get started", check = FALSE,
      open_vignette = FALSE
    )
  )

  map <- map_the_package()
  draw_the_map(map)
})
}
