% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fusen_files.R
\name{register_all_to_config}
\alias{register_all_to_config}
\title{Include all existing package files in the config file}
\usage{
register_all_to_config(pkg)
}
\arguments{
\item{pkg}{Path to the package from which to add file to configuration file}
}
\value{
Path to configuration file
}
\description{
Helps transition from non-fusen packages
}
\examples{
\dontrun{
# Usually run this one inside the current project
register_all_to_config()
}
}
