% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flat_template.R
\name{add_flat_template}
\alias{add_flat_template}
\title{Add flat Rmd file that drives package development}
\usage{
add_flat_template(
  template = c("full", "minimal", "additional", "teaching", "dev_history"),
  pkg = ".",
  dev_dir = "dev",
  flat_name = template,
  overwrite = FALSE,
  open = TRUE
)
}
\arguments{
\item{template}{Name of the template to use. See details.}

\item{pkg}{Path where to save file}

\item{dev_dir}{Name of directory for development Rmarkdown files. Default to "dev".}

\item{flat_name}{Name of the file to write in dev.}

\item{overwrite}{Whether to overwrite existing flat Rmd template file with same name}

\item{open}{Logical. Whether to open file after creation}
}
\value{
Create flat Rmd file(s) and return its (their) path
}
\description{
Add flat Rmd file that drives package development
}
\details{
Choose \code{template} among the different templates available:
\itemize{
\item "full": the full template with a reproducible package to inflate directly. Default.
\item "minimal": Minimal template to start a new package when you already know {fusen}.
\item "additional": Template for an additional vignette, thus additional functions.
\item "teaching": Template with a reproducible package, simpler than "full", but everything to
teach the minimal structure of a package.
}

Abbreviated names can also be used for the different templates. For example "add"
instead of additional.
}
\examples{
# Create a new project
tmpdir <- tempdir()
dummypackage <- file.path(tmpdir, "dummypackage")
dir.create(dummypackage)

# Add
add_flat_template(template = "teaching", pkg = dummypackage)
# Delete dummy package
unlink(dummypackage, recursive = TRUE)

# For classical use in your package
\dontrun{
# first time ever using 'fusen'
add_flat_template("full") 

# first time in your new package
add_flat_template("minimal")

# add new flat file for new functions
add_flat_template("add")
#'}
}
