% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content_parsedtbl.R
\name{combine_parsed}
\alias{combine_parsed}
\title{Combine a parsed tbl Rmd / Qmd file into a new file}
\usage{
combine_parsed(parsed_tbl, output_file)
}
\arguments{
\item{parsed_tbl}{A tibble with 6 columns as issued by \code{get_rmd_parts()}}

\item{output_file}{A Rmd / Qmd file path to write the new content}
}
\value{
The content of a Rmd / Qmd file and
the resulting file if output_file is provided.
}
\description{
Combine a parsed tbl Rmd / Qmd file into a new file
}
