% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{create_vignette}
\alias{create_vignette}
\title{Create vignette}
\usage{
create_vignette(parsed_tbl, pkg, name)
}
\arguments{
\item{parsed_tbl}{tibble of a parsed Rmd}

\item{pkg}{Path to package}

\item{name}{Name of the resulting vignette}
}
\description{
Create vignette
}
