% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate_all.R
\name{inflate_all}
\alias{inflate_all}
\title{Inflate all your flat files}
\usage{
inflate_all(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}
}
\value{
side effect. Inflates all your flat files that can be inflated.
}
\description{
Inflate all the flat files stored in dev/ and starting with "flat_"
}
\examples{
\dontrun{
dummypackage <- tempfile("inflateall")
dir.create(dummypackage)
fill_description(pkg = dummypackage, fields = list(Title = "Dummy Package"))
dev_file <-
  (add_minimal(
    pkg = dummypackage,
    overwrite = TRUE,
    open = FALSE
  ))

usethis::with_project(dummypackage, {
  # if you are starting from a brand new package, inflate_all() will crash
  # it's because of the absence of a fusen config file
  # inflate_all()

  # you need to inflate manually your flat file first
  inflate(
    pkg = dummypackage,
    flat_file = flat_file,
    vignette_name = "Get started",
    check = FALSE,
    open_vignette = FALSE,
    document = TRUE,
    overwrite = "yes"
  )

  # your config file has been created
  config_yml_ref <-
    yaml::read_yaml(getOption("fusen.config_file", default = "dev/config_fusen.yaml"))

  # now you can run inflate_all()
  inflate_all()

  # Let's deprecate our flat file : inflate_all() won't like it and will tell you
  config_yml_deprecated <- config_yml_ref
  config_yml_deprecated[[1]][["state"]] <- "deprecated"
  yaml::write_yaml(config_yml_deprecated, file = "dev/config_fusen.yaml")
  inflate_all()

  # If you add a new flat file, absent from config_fusen.yml, inflate_all() will warn you
  yaml::write_yaml(config_yml_ref, file = "dev/config_fusen.yaml")
  flat_file2 <-
    gsub(
      x = flat_file,
      pattern = "flat_minimal.Rmd",
      replacement = "flat_minimal_2.Rmd"
    )
  file.copy(
    from = flat_file,
    to = flat_file2,
    overwrite = TRUE
  )
  inflate_all()
  unlink(flat_file2)

  # Let's remove the inflate parameters from config_fusen.yml
  # If you used fusen and inflate() prior to v0.5.0.9001 your config_fusen.yml does not contain
  # the inflate parameters. You must inflate it again before being able to use inflate_all()
  config_yml_no_inflate_params <- config_yml_ref
  config_yml_no_inflate_params[[1]][["inflate"]] <- NULL
  yaml::write_yaml(config_yml_no_inflate_params, file = "dev/config_fusen.yaml")
  # inflate_all() will raise an error
  inflate_all()


  # Let's add a flat file in in config_fusen.yml not present in dev/
  config_yml_file_absent_in_dev <- config_yml_ref
  config_yml_file_absent_in_dev[["missing_file.Rmd"]] <-
    config_yml_file_absent_in_dev[[1]]
  yaml::write_yaml(config_yml_file_absent_in_dev, file = "dev/config_fusen.yaml")
  # inflate_all() will raise an error
  inflate_all()
})
unlink(dummypackage, recursive = TRUE)
}
}
